commit : encheres

See merge request eni-encheres1/eni-encheres!79

+++ src/main/java/com/example/encheres/bll/ArticleVenduService.java
@@ -12,9 +12,9 @@ public interface ArticleVenduService {
 	ArticleVendu lectureArticleVendu(int noArticleVendu);
 	List<ArticleVendu> findAll();
 	void modifierArticleVendu(ArticleVendu articleVendu);
-	void modifierArticleVenduPrixVente(int noArticleVendu, float prixVente);
+	void modifierArticleVenduPrixVente(int noArticleVendu, int prixVente);
 	void supprimerArticleVendu(int articleVendu);
 	void createArticleWithRetrait(ArticleVendu articleVendu, Retrait adresse, Utilisateur user);
-	void encherirArticle(int noArticleVendu, float proposition, Utilisateur user);
+	void encherirArticle(int noArticleVendu, int proposition, Utilisateur user);
 	List<ArticleVendu> findAllComplexe(int requete,  String nomArticle, int noCategorie, int noUtilisateurVendeur, int noUtilisateurAcheteur);
 }
